+++ src/main/java/step2/domain/IllegalCountException.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 class IllegalCountException extends IllegalArgumentException {
 }
