+++ src/main/java/step2/domain/MoneyArithmeticException.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 public class MoneyArithmeticException extends ArithmeticException {
     public MoneyArithmeticException(String message) {
