+++ src/main/java/step2/domain/NotEnoughNumberPoolSizeException.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 class NotEnoughNumberPoolSizeException extends IllegalArgumentException {
 }
