+++ src/main/java/step2/domain/OutOfMatchingBoundaryException.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 class OutOfMatchingBoundaryException extends RuntimeException {
 }
