+++ src/main/java/step2/domain/Playslip.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.ArrayList;
 import java.util.Collections;
